Decompose dm-tool BookReader.tsx into 7 sub-modules
Second slice of the C-e BookBrowser/BookReader decomposition. Same
pattern as e1 — strict mechanical extraction, zero behaviour change.

`apps/dm-tool/src/features/book-browser/BookReader.tsx` was 1154 LOC.
After this slice it's ~580 LOC. Extracted pieces live under a new
`BookReader/` directory:

- `BookReader/types.ts` — `OutlineNode`, `TaggedOutlineNode`,
  `ReaderProps`, `DocSlot`
- `BookReader/zoom.ts` — `ZoomPreset`, `ZOOM_PRESETS`, `resolveScale`,
  `loadZoom`, `saveZoom`, `cycleZoom`
- `BookReader/scroll.ts` — `scrollKey`, `loadScroll`, `saveScroll`,
  `PAGE_GAP`, `SEPARATOR_HEIGHT` (the layout constants paired with
  the scroll-position storage helpers since both deal with page
  positioning)
- `BookReader/MultiDocPageList.tsx` — multi-doc page list +
  separator rendering
- `BookReader/PageSlot.tsx` — per-page canvas + text-layer renderer
  with IntersectionObserver-driven lazy loading
- `BookReader/TocTree.tsx` — `TocTree` + `TocNode` + `cleanTocTitle`
  (the Paizo-bookmark cleaner) + the exported `tagNodes` helper used
  by main when building the combined outline
- `BookReader/PageIndicator.tsx` — toolbar page indicator + jump-to-
  page input

`BookReader.tsx` keeps just the root component with its slot lifecycle,
TOC + scroll coordination, and the toolbar render.

Cumulative C-e progress: BookBrowser 1128 → ~470 LOC; BookReader
1154 → ~580 LOC. Bucket C-e of the cleanup audit is complete.

Bucket C-d (test coverage drive) closes out C.
