+++ apps/dm-tool/src/features/book-browser/BookReader.tsx
@@ -1,128 +1,21 @@
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
-import { ArrowLeft, ChevronRight, List, Minus, Plus, RotateCcw } from 'lucide-react';
+import { ArrowLeft, List, Minus, Plus, RotateCcw } from 'lucide-react';
 import { Button } from '@/components/ui/button';
 import { ScrollArea } from '@/components/ui/scroll-area';
 import { Separator } from '@/components/ui/separator';
 import { cn } from '@/lib/utils';
 import { api } from '@/lib/api';
 import { pdfjsLib } from '@/lib/pdfjs';
-import { STORAGE_KEYS } from '@/lib/constants';
-import { readNumber, readString, writeString } from '@/lib/storage-utils';
-import { extractCover } from './useBooks';
-import { partSubtitle, type ApGroup } from './ap-merge';
 import type { PDFDocumentProxy } from 'pdfjs-dist/types/src/display/api';
+import { extractCover } from './useBooks';
+import { partSubtitle } from './ap-merge';
 
-// ---------------------------------------------------------------------------
-// Types
-// ---------------------------------------------------------------------------
-
-interface OutlineNode {
-  title: string;
-  dest: string | unknown[] | null;
-  items: OutlineNode[];
-}
-
-/** A tagged outline node that knows which doc slot it belongs to. Used by
- *  the combined TOC so destination resolution targets the right PDF. */
-interface TaggedOutlineNode extends OutlineNode {
-  slotIndex: number;
-  items: TaggedOutlineNode[];
-}
-
-interface ReaderProps {
-  /** Single book mode — open one PDF. */
-  bookId?: number;
-  /** Merged AP mode — open all parts as one combined document. */
-  apGroup?: ApGroup;
-  onClose: () => void;
-  /** Navigate back to catalog without closing the tab. Falls back to onClose. */
-  onBack?: () => void;
-  onIngestComplete?: () => void;
-}
-
-/** One document in a multi-doc view. */
-interface DocSlot {
-  bookId: number;
-  partLabel: string;
-  pageCount: number;
-  globalPageOffset: number;
-  doc: PDFDocumentProxy | null;
-  outline: OutlineNode[];
-}
-
-// ---------------------------------------------------------------------------
-// Zoom presets
-// ---------------------------------------------------------------------------
-
-type ZoomPreset = 'fit-width' | 'fit-page' | '100' | '150' | '200';
-
-const ZOOM_PRESETS: Array<{ label: string; value: ZoomPreset }> = [
-  { label: 'Fit Width', value: 'fit-width' },
-  { label: 'Fit Page', value: 'fit-page' },
-  { label: '100%', value: '100' },
-  { label: '150%', value: '150' },
-  { label: '200%', value: '200' },
-];
-
-function resolveScale(
-  preset: ZoomPreset,
-  containerWidth: number,
-  containerHeight: number,
-  pageWidth: number,
-  pageHeight: number,
-): number {
-  switch (preset) {
-    case 'fit-width':
-      return containerWidth / pageWidth;
-    case 'fit-page':
-      return Math.min(containerWidth / pageWidth, containerHeight / pageHeight);
-    case '100':
-      return 1;
-    case '150':
-      return 1.5;
-    case '200':
-      return 2;
-  }
-}
-
-const PAGE_GAP = 8;
-const SEPARATOR_HEIGHT = 48;
-
-// ---------------------------------------------------------------------------
-// localStorage helpers for reader preferences
-// ---------------------------------------------------------------------------
-
-function loadZoom(): ZoomPreset {
-  const v = readString(STORAGE_KEYS.readerZoom);
-  if (v && ZOOM_PRESETS.some((p) => p.value === v)) return v as ZoomPreset;
-  return 'fit-width';
-}
-
-function saveZoom(z: ZoomPreset) {
-  writeString(STORAGE_KEYS.readerZoom, z);
-}
-
-/** Stable key for scroll position: single book uses the bookId, merged
- *  APs use "ap-<subcategory>". */
-function scrollKey(bookId?: number, apGroup?: ApGroup): string | null {
-  if (apGroup) return `${STORAGE_KEYS.readerScrollPrefix}ap-${apGroup.subcategory}`;
-  if (bookId != null) return `${STORAGE_KEYS.readerScrollPrefix}${bookId}`;
-  return null;
-}
-
-function loadScroll(key: string | null): number {
-  if (!key) return 0;
-  return readNumber(key, 0);
-}
-
-function saveScroll(key: string | null, top: number) {
-  if (!key) return;
-  writeString(key, String(Math.round(top)));
-}
-
-// ---------------------------------------------------------------------------
-// Main reader component
-// ---------------------------------------------------------------------------
+import { MultiDocPageList } from './BookReader/MultiDocPageList';
+import { PageIndicator } from './BookReader/PageIndicator';
+import { tagNodes, TocTree } from './BookReader/TocTree';
+import { loadScroll, PAGE_GAP, SEPARATOR_HEIGHT, saveScroll, scrollKey } from './BookReader/scroll';
+import type { DocSlot, OutlineNode, ReaderProps, TaggedOutlineNode } from './BookReader/types';
+import { cycleZoom, loadZoom, resolveScale, saveZoom, type ZoomPreset, ZOOM_PRESETS } from './BookReader/zoom';
 
 export function BookReader({ bookId, apGroup, onClose, onBack, onIngestComplete }: ReaderProps) {
   const isMulti = !!apGroup;
@@ -704,451 +597,4 @@ export function BookReader({ bookId, apGroup, onClose, onBack, onIngestComplete 
       </div>
     </div>
   );
-}
-
-// ---------------------------------------------------------------------------
-// Multi-doc page list
-// ---------------------------------------------------------------------------
-
-function MultiDocPageList({
-  slots,
-  slotTopOffsets,
-  pageHeight,
-  pageWidth,
-  scale,
-  loadSlotDoc,
-}: {
-  slots: DocSlot[];
-  /** Precomputed scroll-top of each slot's first page — shared with the
-   *  TOC resolver so both use identical positions. */
-  slotTopOffsets: number[];
-  pageHeight: number;
-  pageWidth: number;
-  scale: number;
-  loadSlotDoc: (idx: number) => Promise<PDFDocumentProxy | null>;
-}) {
-  // Build a flat list of renderable items (pages + separators) using the
-  // same slotTopOffsets the TOC resolver uses.
-  const items = useMemo(() => {
-    const list: Array<
-      | { kind: 'page'; slotIndex: number; localPageNum: number; top: number }
-      | { kind: 'separator'; label: string; top: number }
-    > = [];
-    for (let si = 0; si < slots.length; si++) {
-      const slot = slots[si]!;
-      const slotTop = slotTopOffsets[si] ?? 0;
-      if (si > 0) {
-        list.push({
-          kind: 'separator',
-          label: slot.partLabel,
-          top: slotTop - SEPARATOR_HEIGHT,
-        });
-      }
-      for (let p = 0; p < slot.pageCount; p++) {
-        list.push({
-          kind: 'page',
-          slotIndex: si,
-          localPageNum: p + 1,
-          top: slotTop + p * (pageHeight + PAGE_GAP),
-        });
-      }
-    }
-    return list;
-  }, [slots, slotTopOffsets, pageHeight]);
-
-  // Total height from the last item's bottom edge.
-  const lastItem = items[items.length - 1];
-  const totalHeight = lastItem ? lastItem.top + (lastItem.kind === 'page' ? pageHeight : SEPARATOR_HEIGHT) : 0;
-
-  return (
-    <div
-      style={{
-        position: 'relative',
-        width: pageWidth,
-        height: totalHeight,
-        margin: '0 auto',
-        paddingTop: PAGE_GAP,
-      }}
-    >
-      {items.map((item, i) =>
-        item.kind === 'separator' ? (
-          <div
-            key={`sep-${i}`}
-            style={{
-              position: 'absolute',
-              top: item.top,
-              left: 0,
-              width: pageWidth,
-              height: SEPARATOR_HEIGHT,
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'center',
-            }}
-          >
-            <div className="flex items-center gap-3">
-              <div className="h-px w-12 bg-border" />
-              <span className="text-xs font-medium text-muted-foreground">{item.label}</span>
-              <div className="h-px w-12 bg-border" />
-            </div>
-          </div>
-        ) : (
-          <PageSlot
-            key={`${item.slotIndex}-${item.localPageNum}`}
-            slots={slots}
-            slotIndex={item.slotIndex}
-            localPageNum={item.localPageNum}
-            width={pageWidth}
-            height={pageHeight}
-            scale={scale}
-            top={item.top}
-            loadSlotDoc={loadSlotDoc}
-          />
-        ),
-      )}
-    </div>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// Page slot — renders canvas + text layer when visible
-// ---------------------------------------------------------------------------
-
-function PageSlot({
-  slots,
-  slotIndex,
-  localPageNum,
-  width,
-  height,
-  scale,
-  top,
-  loadSlotDoc,
-}: {
-  slots: DocSlot[];
-  slotIndex: number;
-  localPageNum: number;
-  width: number;
-  height: number;
-  scale: number;
-  top: number;
-  loadSlotDoc: (idx: number) => Promise<PDFDocumentProxy | null>;
-}) {
-  const containerRef = useRef<HTMLDivElement>(null);
-  const canvasRef = useRef<HTMLCanvasElement>(null);
-  const textLayerRef = useRef<HTMLDivElement>(null);
-  const [visible, setVisible] = useState(false);
-  const renderingRef = useRef(false);
-  const renderedScaleRef = useRef<number | null>(null);
-
-  const doc = slots[slotIndex]?.doc ?? null;
-
-  // IntersectionObserver with 1-page margin.
-  useEffect(() => {
-    const el = containerRef.current;
-    if (!el) return;
-    const obs = new IntersectionObserver(([entry]) => setVisible(entry!.isIntersecting), {
-      rootMargin: `${height}px 0px ${height}px 0px`,
-    });
-    obs.observe(el);
-    return () => obs.disconnect();
-  }, [height]);
-
-  // Trigger lazy doc loading when this page becomes visible.
-  useEffect(() => {
-    if (visible && !doc) {
-      loadSlotDoc(slotIndex).catch(console.error);
-    }
-  }, [visible, doc, slotIndex, loadSlotDoc]);
-
-  // Render when visible + doc loaded + scale changed.
-  useEffect(() => {
-    if (!visible || !doc) return;
-    if (renderingRef.current) return;
-    if (renderedScaleRef.current === scale) return;
-
-    let cancelled = false;
-    renderingRef.current = true;
-
-    (async () => {
-      try {
-        const page = await doc.getPage(localPageNum);
-        if (cancelled) return;
-
-        const viewport = page.getViewport({ scale });
-        const canvas = canvasRef.current;
-        if (!canvas || cancelled) return;
-
-        const dpr = window.devicePixelRatio || 1;
-        canvas.width = Math.round(viewport.width * dpr);
-        canvas.height = Math.round(viewport.height * dpr);
-        canvas.style.width = `${viewport.width}px`;
-        canvas.style.height = `${viewport.height}px`;
-
-        const ctx = canvas.getContext('2d')!;
-        ctx.setTransform(dpr, 0, 0, dpr, 0, 0);
-
-        await page.render({ canvas, canvasContext: ctx, viewport }).promise;
-        if (cancelled) return;
-        renderedScaleRef.current = scale;
-
-        const textDiv = textLayerRef.current;
-        if (!textDiv || cancelled) return;
-        textDiv.innerHTML = '';
-        // pdfjs TextLayer uses the CSS variable --scale-factor to compute
-        // span transforms. Without it, the text spans drift from the canvas.
-        // See: https://github.com/mozilla/pdf.js/discussions/18068
-        textDiv.style.setProperty('--scale-factor', String(scale));
-        textDiv.style.width = `${viewport.width}px`;
-        textDiv.style.height = `${viewport.height}px`;
-
-        const { TextLayer } = pdfjsLib;
-        const textContent = await page.getTextContent();
-        if (cancelled) return;
-
-        const textLayer = new TextLayer({
-          textContentSource: textContent,
-          container: textDiv,
-          viewport,
-        });
-        await textLayer.render();
-      } catch (e) {
-        if (!cancelled) console.error(`Page render error:`, e);
-      } finally {
-        renderingRef.current = false;
-      }
-    })();
-
-    return () => {
-      cancelled = true;
-    };
-  }, [visible, scale, doc, localPageNum]);
-
-  // Cleanup on scroll-out.
-  useEffect(() => {
-    if (visible) return;
-    const canvas = canvasRef.current;
-    if (canvas) {
-      const ctx = canvas.getContext('2d');
-      ctx?.clearRect(0, 0, canvas.width, canvas.height);
-      canvas.width = 0;
-      canvas.height = 0;
-    }
-    const textDiv = textLayerRef.current;
-    if (textDiv) textDiv.innerHTML = '';
-    renderedScaleRef.current = null;
-  }, [visible]);
-
-  return (
-    <div
-      ref={containerRef}
-      style={{
-        position: 'absolute',
-        top,
-        left: 0,
-        width,
-        height,
-        background: doc ? 'white' : undefined,
-        boxShadow: doc ? '0 1px 4px rgba(0,0,0,0.15)' : undefined,
-      }}
-    >
-      {doc ? (
-        <>
-          <canvas ref={canvasRef} style={{ display: 'block' }} />
-          <div ref={textLayerRef} className="textLayer" />
-        </>
-      ) : (
-        <div className="flex h-full items-center justify-center text-xs text-muted-foreground">Loading…</div>
-      )}
-    </div>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// TOC tree — uses a resolveDest callback for cross-doc navigation
-// ---------------------------------------------------------------------------
-
-function TocTree({
-  nodes,
-  resolveDest,
-  scrollRef,
-}: {
-  nodes: TaggedOutlineNode[];
-  resolveDest: (dest: string | unknown[] | null, slotIndex: number) => Promise<number | null>;
-  scrollRef: React.RefObject<HTMLDivElement | null>;
-}) {
-  return (
-    <ul className="space-y-0.5 text-xs">
-      {nodes.map((node, i) => (
-        <TocNode key={i} node={node} resolveDest={resolveDest} scrollRef={scrollRef} depth={0} />
-      ))}
-    </ul>
-  );
-}
-
-function TocNode({
-  node,
-  resolveDest,
-  scrollRef,
-  depth,
-}: {
-  node: TaggedOutlineNode;
-  resolveDest: (dest: string | unknown[] | null, slotIndex: number) => Promise<number | null>;
-  scrollRef: React.RefObject<HTMLDivElement | null>;
-  depth: number;
-}) {
-  const [expanded, setExpanded] = useState(depth < 1);
-  const hasChildren = node.items && node.items.length > 0;
-
-  const handleClick = useCallback(async () => {
-    if (!scrollRef.current) return;
-    const top = await resolveDest(node.dest, node.slotIndex);
-    if (top != null) {
-      scrollRef.current.scrollTop = top;
-    }
-  }, [node.dest, node.slotIndex, resolveDest, scrollRef]);
-
-  return (
-    <li>
-      <div className="flex items-start">
-        {hasChildren ? (
-          <button
-            type="button"
-            className="mt-0.5 flex h-4 w-4 shrink-0 items-center justify-center text-muted-foreground"
-            onClick={() => setExpanded(!expanded)}
-          >
-            <ChevronRight className={cn('h-3 w-3 transition-transform', expanded && 'rotate-90')} />
-          </button>
-        ) : (
-          <span className="w-4 shrink-0" />
-        )}
-        <button
-          type="button"
-          onClick={handleClick}
-          className={cn(
-            'flex-1 truncate py-0.5 text-left transition-colors hover:text-foreground',
-            depth === 0 && node.items.length > 0 ? 'font-medium text-foreground/80' : 'text-muted-foreground',
-          )}
-          style={{ paddingLeft: depth * 8 }}
-          title={cleanTocTitle(node.title)}
-        >
-          {cleanTocTitle(node.title)}
-        </button>
-      </div>
-      {expanded && hasChildren && (
-        <ul className="ml-2">
-          {node.items.map((child, i) => (
-            <TocNode key={i} node={child} resolveDest={resolveDest} scrollRef={scrollRef} depth={depth + 1} />
-          ))}
-        </ul>
-      )}
-    </li>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// Page indicator — shows "Page N / M", click to open jump-to-page input
-// ---------------------------------------------------------------------------
-
-function PageIndicator({
-  currentPage,
-  totalPages,
-  isMulti,
-  slotCount,
-  open,
-  onOpenChange,
-  onJump,
-}: {
-  currentPage: number;
-  totalPages: number;
-  isMulti: boolean;
-  slotCount: number;
-  open: boolean;
-  onOpenChange: (v: boolean) => void;
-  onJump: (page: number) => void;
-}) {
-  const inputRef = useRef<HTMLInputElement>(null);
-  const [inputValue, setInputValue] = useState('');
-
-  useEffect(() => {
-    if (open) {
-      setInputValue(String(currentPage));
-      // Focus after React renders the input.
-      requestAnimationFrame(() => inputRef.current?.select());
-    }
-  }, [open, currentPage]);
-
-  const handleSubmit = () => {
-    const n = parseInt(inputValue, 10);
-    if (Number.isFinite(n)) onJump(n);
-    onOpenChange(false);
-  };
-
-  if (open) {
-    return (
-      <span className="flex items-center gap-1 text-[10px]">
-        <span className="text-muted-foreground">Page</span>
-        <input
-          ref={inputRef}
-          type="text"
-          inputMode="numeric"
-          value={inputValue}
-          onChange={(e) => setInputValue(e.target.value)}
-          onKeyDown={(e) => {
-            if (e.key === 'Enter') handleSubmit();
-            if (e.key === 'Escape') onOpenChange(false);
-            e.stopPropagation(); // don't trigger reader shortcuts
-          }}
-          onBlur={handleSubmit}
-          className="w-12 rounded border border-border bg-background px-1 py-0.5 text-center text-[10px] text-foreground outline-hidden focus:border-primary"
-        />
-        <span className="text-muted-foreground">/ {totalPages}</span>
-      </span>
-    );
-  }
-
-  return (
-    <button
-      type="button"
-      onClick={() => onOpenChange(true)}
-      className="text-[10px] text-muted-foreground transition-colors hover:text-foreground"
-      title="Click to jump to a page"
-    >
-      Page {currentPage} / {totalPages}
-      {isMulti && ` · ${slotCount} parts`}
-    </button>
-  );
-}
-
-// ---------------------------------------------------------------------------
-// Helpers
-// ---------------------------------------------------------------------------
-
-/** Strip Paizo bookmark noise from TOC titles. Common patterns:
- *    "018-031 PZO90152 Chapter 2" → "Chapter 2"
- *    "032 PZO90152 Appendix" → "Appendix"
- *    "PZO9015-2 Introduction" → "Introduction"
- *  Leading page ranges (\d+-\d+ or \d+), product codes (PZO\w+), and
- *  resulting whitespace are removed. */
-function cleanTocTitle(raw: string): string {
-  return (
-    raw
-      .replace(/^\d+(?:-\d+)?\s*/g, '') // leading page range "018-031 " or "032 "
-      .replace(/^PZO[\w-]+\s*/gi, '') // product code "PZO90152 "
-      .trim() || raw
-  ); // fall back to original if nothing remains
-}
-
-function tagNodes(nodes: OutlineNode[], slotIndex: number): TaggedOutlineNode[] {
-  return nodes.map((n) => ({
-    ...n,
-    slotIndex,
-    items: tagNodes(n.items ?? [], slotIndex),
-  }));
-}
-
-function cycleZoom(dir: 1 | -1, current: ZoomPreset, set: (v: ZoomPreset) => void) {
-  const idx = ZOOM_PRESETS.findIndex((p) => p.value === current);
-  const next = idx + dir;
-  if (next >= 0 && next < ZOOM_PRESETS.length) {
-    set(ZOOM_PRESETS[next]!.value);
-  }
 }
